+++ lib/httpOperationResponse.ts
@@ -5,7 +5,7 @@ import { WebResource } from "./webResource";
 
 /**
  * Wrapper object for http request and response. Deserialized object is stored in
- * the `body` property.
+ * the `bodyAsJson` property when the response body is received in JSON.
  * @class
  * Initializes a new instance of the HttpOperationResponse class.
  * @constructor
@@ -16,13 +16,9 @@ export class HttpOperationResponse {
    */
   request: WebResource;
   /**
-   * The raw response
+   * The raw response. Please use the response directly when the response body is a ReadableStream.
    */
   response: Response;
-  /**
-   * The response body as a readable stream
-   */
-  bodyAsStream: ReadableStream | null;
   /**
    * The response body as text (string format)
    */
@@ -32,7 +28,7 @@ export class HttpOperationResponse {
    */
   bodyAsJson: { [key: string]: any } | Array<any> | string | number | boolean | null | void;
 
-  constructor(request: WebResource, response: Response, body: ReadableStream | null) {
+  constructor(request: WebResource, response: Response) {
     /**
      * Reference to the original request object.
      * [WebResource] object.
@@ -46,12 +42,7 @@ export class HttpOperationResponse {
      * @type {object}
      */
     this.response = response;
-
-    /**
-     * The response object.
-     * @type {object}
-     */
-    this.bodyAsStream = body;
+    /* tslint:disable:no-null-keyword */
     this.bodyAsText = null;
     this.bodyAsJson = null;
   }